+++ NEWS.md
@@ -3,6 +3,7 @@
     * Allow sf points for sampling metrics
     * Make sure x-y coordinates are used in sample_lsm()
     * size argument is not needed for polygons in sample_lsm()
+    * make sure sample_lsm() can use SpatialPolygonsDataFrame
 
 # landscapemetrics 1.2
 * New functions